pyproject.toml
[tool.poetry]
name = "files"
version = "0.1.0"
description = ""
authors = ["Nazhmutdin <gumuevnazhmutdin1248@gmail.com>"]
readme = "README.md"
packages = [
    { include = "app", from = "src" }
]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
dishka = "^1.4.1"
fastapi = "^0.115.5"
aiofiles = "^24.1.0"
uvicorn = "^0.32.0"
orjson = "^3.10.11"
python-multipart = "^0.0.17"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pytest-asyncio = "^0.24.0"
faker = "^30.6.0"
certifi = "^2024.8.30"
httpx = "^0.27.2"
pytest-dotenv = "^0.5.2"
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = "./src"
env_files = "./.test.env"
